+++ spec/models/artist.rb
@@ -1,0 +1,7 @@
+class Artist
+  include Mongoid::Document
+
+  field :name
+
+  has_many :artworks
+end